+++ CHANGELOG.md
@@ -9,3 +9,10 @@ All notable changes to the "ipython" extension will be documented in this file.
 - Install various tunable delays so code can be sent to the terminal (rather
   than using the `%run` magic command)
 - Option to have a different terminal per file.
+
+## 2022-06-17
+
+- Forked from [shaperilio/vscode-ipython](https://github.com/shaperilio/vscode-ipython))
+- Introduce cell separator (horizontal rule)
+- Option to run whole file without %run command
+- Elaborate multi-line code execution